requirements.txt
torch
torchvision
transformers
timm
opencv-python
numpy
pandas
matplotlib
seaborn
scikit-learn
streamlit
pillow
tqdm
streamlit-webrtc
flask
pymongo
werkzeug
python-dotenv
sentencepiece
fpdf2
